Add files via upload
I created this file and I am adding it to a Flask project I am working on.  I have a member of the team works with APC and every time he needs one added to ISE he asks me to add it for him.  Since they are all setup for the same policy set and rules I wrote this so on the webapp all he needs to do is add the IP address.  This is a bit sterilized from what I actually have.  I have MY_ATH.py that has all my host and user info for ERS on ISE.